Remove the IOCCCWinnersDB class.
Completely useless.
